+++ src/test/java/com/uber/cadence/worker/StickyWorkerTest.java
@@ -34,6 +34,8 @@ import com.uber.cadence.client.WorkflowOptions;
 import com.uber.cadence.internal.metrics.MetricsTag;
 import com.uber.cadence.internal.metrics.MetricsType;
 import com.uber.cadence.internal.replay.DeciderCache;
+import com.uber.cadence.serviceclient.IWorkflowService;
+import com.uber.cadence.serviceclient.WorkflowServiceTChannel;
 import com.uber.cadence.testing.TestEnvironmentOptions;
 import com.uber.cadence.testing.TestWorkflowEnvironment;
 import com.uber.cadence.workflow.Async;
@@ -56,24 +58,19 @@ import java.util.List;
 import java.util.Map;
 import java.util.Queue;
 import java.util.Random;
-import org.junit.Assert;
-import org.junit.Ignore;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.*;
 import org.junit.rules.TestName;
+import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-// @RunWith(Parameterized.class)
-@Ignore
+@RunWith(Parameterized.class)
 public class StickyWorkerTest {
   public static final String DOMAIN = "UnitTest";
 
-  // TODO: Enable for docker as soon as the server commit
-  // a36c84991664571636d37a3826b282ddbdbd2402 is released
-  private static final boolean skipDockerService = true;
-  //      Boolean.parseBoolean(System.getenv("SKIP_DOCKER_SERVICE"));
+  private static final boolean skipDockerService =
+      Boolean.parseBoolean(System.getenv("SKIP_DOCKER_SERVICE"));
 
   @Parameterized.Parameter public boolean useExternalService;
 
@@ -91,6 +88,15 @@ public class StickyWorkerTest {
 
   @Rule public TestName testName = new TestName();
 
+  private IWorkflowService service;
+
+  @Before
+  public void setUp() {
+    if (testType.equals("Docker") && service == null) {
+      service = new WorkflowServiceTChannel();
+    }
+  }
+
   @Test
   public void whenStickyIsEnabledThenTheWorkflowIsCachedSignals() throws Exception {
     // Arrange
@@ -104,7 +110,10 @@ public class StickyWorkerTest {
 
     TestEnvironmentWrapper wrapper =
         new TestEnvironmentWrapper(
-            new Worker.FactoryOptions.Builder().setMetricScope(scope).build());
+            new Worker.FactoryOptions.Builder()
+                .setDisableStickyExecution(false)
+                .setMetricScope(scope)
+                .build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName, new WorkerOptions.Builder().build());
     worker.registerWorkflowImplementationTypes(GreetingSignalWorkflowImpl.class);
@@ -159,7 +168,10 @@ public class StickyWorkerTest {
 
     TestEnvironmentWrapper wrapper =
         new TestEnvironmentWrapper(
-            new Worker.FactoryOptions.Builder().setMetricScope(scope).build());
+            new Worker.FactoryOptions.Builder()
+                .setDisableStickyExecution(false)
+                .setMetricScope(scope)
+                .build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName, new WorkerOptions.Builder().build());
     worker.registerWorkflowImplementationTypes(ActivitiesWorkflowImpl.class);
@@ -214,7 +226,10 @@ public class StickyWorkerTest {
 
     TestEnvironmentWrapper wrapper =
         new TestEnvironmentWrapper(
-            new Worker.FactoryOptions.Builder().setMetricScope(scope).build());
+            new Worker.FactoryOptions.Builder()
+                .setDisableStickyExecution(false)
+                .setMetricScope(scope)
+                .build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName, new WorkerOptions.Builder().build());
     worker.registerWorkflowImplementationTypes(
@@ -262,7 +277,10 @@ public class StickyWorkerTest {
 
     TestEnvironmentWrapper wrapper =
         new TestEnvironmentWrapper(
-            new Worker.FactoryOptions.Builder().setMetricScope(scope).build());
+            new Worker.FactoryOptions.Builder()
+                .setDisableStickyExecution(false)
+                .setMetricScope(scope)
+                .build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName, new WorkerOptions.Builder().build());
     worker.registerWorkflowImplementationTypes(TestMutableSideEffectWorkflowImpl.class);
@@ -343,7 +361,8 @@ public class StickyWorkerTest {
     // Arrange
     String taskListName = "evictedStickyTest";
     TestEnvironmentWrapper wrapper =
-        new TestEnvironmentWrapper(new Worker.FactoryOptions.Builder().build());
+        new TestEnvironmentWrapper(
+            new Worker.FactoryOptions.Builder().setDisableStickyExecution(false).build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName);
     worker.registerWorkflowImplementationTypes(GreetingSignalWorkflowImpl.class);
@@ -382,7 +401,8 @@ public class StickyWorkerTest {
     // Arrange
     String taskListName = "queryStickyTest";
     TestEnvironmentWrapper wrapper =
-        new TestEnvironmentWrapper(new Worker.FactoryOptions.Builder().build());
+        new TestEnvironmentWrapper(
+            new Worker.FactoryOptions.Builder().setDisableStickyExecution(false).build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName);
     worker.registerWorkflowImplementationTypes(GreetingSignalWorkflowImpl.class);
@@ -422,7 +442,8 @@ public class StickyWorkerTest {
     // Arrange
     String taskListName = "queryEvictionStickyTest";
     TestEnvironmentWrapper wrapper =
-        new TestEnvironmentWrapper(new Worker.FactoryOptions.Builder().build());
+        new TestEnvironmentWrapper(
+            new Worker.FactoryOptions.Builder().setDisableStickyExecution(false).build());
     Worker.Factory factory = wrapper.getWorkerFactory();
     Worker worker = factory.newWorker(taskListName);
     worker.registerWorkflowImplementationTypes(GreetingSignalWorkflowImpl.class);
@@ -467,12 +488,18 @@ public class StickyWorkerTest {
 
     public TestEnvironmentWrapper(Worker.FactoryOptions options) {
       if (options == null) {
-        options = new Worker.FactoryOptions.Builder().build();
+        options = new Worker.FactoryOptions.Builder().setDisableStickyExecution(false).build();
       }
-      factory = new Worker.Factory(DOMAIN, options);
-      TestEnvironmentOptions testOptions =
-          new TestEnvironmentOptions.Builder().setDomain(DOMAIN).setFactoryOptions(options).build();
-      testEnv = TestWorkflowEnvironment.newInstance(testOptions);
+      if (useExternalService) {
+        factory = new Worker.Factory(service, DOMAIN, options);
+      } else {
+        TestEnvironmentOptions testOptions =
+            new TestEnvironmentOptions.Builder()
+                .setDomain(DOMAIN)
+                .setFactoryOptions(options)
+                .build();
+        testEnv = TestWorkflowEnvironment.newInstance(testOptions);
+      }
     }
 
     private Worker.Factory getWorkerFactory() {
@@ -480,12 +507,17 @@ public class StickyWorkerTest {
     }
 
     private WorkflowClient getWorkflowClient() {
-      return useExternalService ? WorkflowClient.newInstance(DOMAIN) : testEnv.newWorkflowClient();
+      return useExternalService
+          ? WorkflowClient.newInstance(service, DOMAIN)
+          : testEnv.newWorkflowClient();
     }
 
     private void close() {
-      factory.shutdown(Duration.ofSeconds(1));
-      testEnv.close();
+      if (useExternalService) {
+        factory.shutdown(Duration.ofSeconds(1));
+      } else {
+        testEnv.close();
+      }
     }
   }