+++ CHANGELOG.md
@@ -11,3 +11,11 @@
 - Side effects, mutable side effects, random uuid and workflow getVersion support.
 - Activity heartbeat throttling.
 - Deterministic retry of failed operation.
+
+## v2.1.2
+- Requires minimum server release v0.4.0
+- Introduced WorkerFactory and FactoryOptions
+- Added sticky workflow execution, which is caching of a workflow object between decisions. It is enabled by default, 
+to disable use FactoryOptions.disableStickyExecution property.
+- Updated Thrift to expose new types of service exceptions: ServiceBusyError, DomainNotActiveError, LimitExceededError
+- Added metric for corrupted signal as well as metrics related to caching and evictions.